% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_spacy_lang_model.R
\name{download_spacy_lang_model}
\alias{download_spacy_lang_model}
\title{Download en_core_web_sm}
\usage{
download_spacy_lang_model()
}
\description{
Download the small English spaCy langauge model.
}
\examples{
download_spacy_lang_model()

}
