% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_triples.R
\name{extract_triples}
\alias{extract_triples}
\title{Extract grammatical triples from a text column in a data frame.}
\usage{
extract_triples()
}
\description{
Extract grammatical triples from a text column in a data frame.
}
\examples{
extract_triples(dataframe, col_name)

}
