% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posextract_initialize.R
\name{posextract_initialize}
\alias{posextract_initialize}
\title{Initialize the posextract virtual environment (required when using posextract)}
\usage{
posextract_initialize()
}
\description{
Initialize the posextract virtual environment (required when using posextract)
}
\examples{
posextract_initialize()

}
