% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posextract_install.R
\name{posextract_install}
\alias{posextract_install}
\title{Install posextract}
\usage{
posextract_install()
}
\description{
Prompts the user to install or posextractr and dependencies. Users have the option to skip.
}
