% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_adj_noun_pairs.R
\name{extract_adj_noun_pairs}
\alias{extract_adj_noun_pairs}
\title{Extract grammatical adjective-noun pairs from a text column in a data frame.}
\usage{
extract_adj_noun_pairs()
}
\description{
Extract grammatical adjective-noun pairs from a text column in a data frame.
}
\examples{
extract_adj_noun_pairs(dataframe, col_name)

}
