% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subj_verb_pairs.R
\name{extract_subj_verb_pairs}
\alias{extract_subj_verb_pairs}
\title{Subject-Verb Pairs}
\usage{
extract_subj_verb_pairs(vect, lemmatize)
}
\description{
Extract subject-verb pairs that grammatically agree with one-another from a text column in a data frame.
}
