% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_triples.R
\name{extract_triples}
\alias{extract_triples}
\title{Triples}
\usage{

data("hansard_random"")

extract_triples(vect, combine_adj = F, lemmatize = F, add_aux = F)

}
\description{
Extract triples that grammatically agree with one-another from a text column in a data frame.
}
