% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_triples.R
\name{extract_triples}
\alias{extract_triples}
\title{Triples}
\usage{
extract_triples()
}
\description{
Extract triples that grammatically agree with one-another from a text column in a data frame.
}
\examples{
extract_triples(dataframe$col_name)

}
