% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posextract_initialize.R
\name{posextract_initialize}
\alias{posextract_initialize}
\title{Initialize posextract}
\usage{
posextract_initialize()
}
\description{
Initialize the posextract Python virtual environment. The environment must be activated to use functions from posextract.
}
