% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posextract_install.R
\name{posextract_install}
\alias{posextract_install}
\title{Install posextractr and dependencies guided by a prompt.}
\usage{
posextract_install()
}
\description{
Install posextractr and dependencies guided by a prompt.
}
\examples{
posextract_install()

}
